[#221] Add first unit test on ideal case
This first test works by sequentially altering the mock data file from local to comp and then back to local mode.
The stdout of the script is used currently to check if the proper behavior is seen. I plan on adding some checks directly on the files itself.
I have setup a fixture to run around tests as a redundancy to be able to reset the mock data properly, though it doesn't actually reset the file yet.

Perhaps a better/additional approach might be to create the local and comp mode snippets separately and use those to compare against the result.