[#209] Initial RoverCommander class, publishing and client request examples tested with CLI script
To start fleshing out this class for now I am building it
while relying on RoverControlCli.py to handle the rate of sending those drive commands.
I haven't implemented checking responses yet but that's next.
Also I'm not sure yet if it's feasible but I intend on trying to abstract the stream of commands
so that developers who use the class need not worry about creating an event loop.
This might only be possible by adding some new requests to the list of rover requests.
However, when the request is made the the rover listener will have to account for the fact that network connection may fail,
so I'm thinking of implementing a polling to the basestation or perhaps even more specifically to the node itself (currently anonymous, might have to change that).
If through polling the CommandListener sees that the connection is lost it will forward an immediate STOP command to the rover MCU.